in got_opentempfd() don't attempt to unlink the file if mkstemp(3) failed

+++ lib/opentemp.c
@@ -34,7 +34,8 @@ got_opentempfd(void)
 		return -1;
 
 	fd = mkstemp(name);
-	unlink(name);
+	if (fd != -1)
+		unlink(name);
 	return fd;
 }
